% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdXdata-package.R
\docType{package}
\name{bdXdata-package}
\alias{bdXdata-package}
\alias{_PACKAGE}
\alias{bdXdata}
\title{bdXdata: Shared data in use at the CRP2A laboratory.}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A collection of shared data in use at the CRP2A laboratory.
}
\details{
\tabular{ll}{
\strong{Package:} \tab bdXdata \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.1.0 \cr
\strong{License:} \tab GPL-3
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/bdXdata}
  \item Report bugs at \url{https://github.com/crp2a/bdXdata/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Brice Lebrun\cr
\email{brice.lebrun@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
