% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BEGe_60cc}
\alias{BEGe_60cc}
\title{BEGe 60cc Calibration Data}
\format{
A \code{\link{list}} of \code{\link{data.frame}} with 8 variables:
\describe{
\item{\code{isotope}}{Measured isotope.}
\item{\code{serie}}{Radioactive decay series to which the considered isotope
belongs.}
\item{\code{energy}}{Theoretical energy of the considered peak.}
\item{\code{bkg}}{Background intensity value (c/ks).}
\item{\code{bkg_error}}{Background error intensity value (c/ks).}
\item{\code{standard_name}}{Name of the standard used.}
\item{\code{standard}}{Standard intensity value (c/ks).}
\item{\code{standard_error}}{Standard error intensity value (c/ks).}
}
}
\usage{
BEGe_60cc
}
\description{
Background and standard intensity values for 60cc BEGe.
}
\seealso{
Other BEGe: 
\code{\link{BEGe_12cc}}
}
\concept{BEGe}
\keyword{datasets}
