% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\docType{data}
\name{clermont}
\alias{clermont}
\title{Clermont Reference Data}
\format{
An object of class \code{data.frame} with 9 rows and 15 columns.
}
\source{
Guérin, G., Mercier, N. & Adamiec, G. (2011). Dose-Rate Conversion Factors:
Update. \emph{Ancient TL}, 29(1): 5-8.

Miallier, D., Guérin, G., Mercier, N., Pilleyre, T. & Sanzelle, S.
(2009). The Clermont Radiometric Reference Rocks: A Convenient Tool
for Dosimetric Purposes. \emph{Ancient TL}, 27(2): 37-44.
}
\usage{
clermont
}
\description{
Clermont Reference Data
}
\seealso{
Other standards: 
\code{\link{std_activity}}
}
\concept{standards}
\keyword{datasets}
