% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{Isotopic Data}
\format{
A \code{\link{data.frame}} with 18 observations of 4 variables:
\describe{
\item{\code{isotope}}{Isotope.}
\item{\code{energy}}{Theoretical energy of the peak (keV).}
\item{\code{mac}}{Mass absorption coefficient (cm2/g).}
\item{\code{proba_emission}}{}
}
}
\usage{
isotopes
}
\description{
Isotopic Data
}
\seealso{
Other references: 
\code{\link{interferences}},
\code{\link{ref_mac}}
}
\concept{references}
\keyword{datasets}
