% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\docType{data}
\name{ref_interfer}
\alias{ref_interfer}
\title{Interferences}
\format{
A \code{\link{data.frame}} with 8 observations of 6 variables:
\describe{
\item{\code{isotope}}{Isotope.}
\item{\code{energy}}{Theoretical energy of the peak (keV).}
\item{\code{target}}{}
\item{\code{target_energy}}{(keV).}
\item{\code{probability}}{}
\item{\code{probability_error}}{}
}
}
\usage{
ref_interfer
}
\description{
Interferences
}
\seealso{
Other references: 
\code{\link{ref_mac}}
}
\concept{references}
\keyword{datasets}
