% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\docType{data}
\name{ref_mac}
\alias{ref_mac}
\title{Mass Absorption Coefficient}
\format{
A \code{\link{data.frame}} with 22 observations of 18 variables:
\describe{
\item{\code{energy}}{Energy (keV).}
\item{\code{hydrogen}}{Mass absorption coefficient of hydrogen (cm2/g).}
\item{\code{carbon}}{Mass absorption coefficient of carbon (cm2/g).}
\item{\code{nitrogen}}{Mass absorption coefficient of nitrogen (cm2/g).}
\item{\code{oxygen}}{Mass absorption coefficient of oxygen (cm2/g).}
\item{\code{fluorine}}{Mass absorption coefficient of fluorine (cm2/g).}
\item{\code{sodium}}{Mass absorption coefficient of sodium (cm2/g).}
\item{\code{magnesium}}{Mass absorption coefficient of magnesium (cm2/g).}
\item{\code{aluminium}}{Mass absorption coefficient of aluminium (cm2/g).}
\item{\code{silicium}}{Mass absorption coefficient of silicium (cm2/g).}
\item{\code{phosphorus}}{Mass absorption coefficient of phosphorus (cm2/g).}
\item{\code{sulfur}}{Mass absorption coefficient of sulfur (cm2/g).}
\item{\code{chlorine}}{Mass absorption coefficient of chlorine (cm2/g).}
\item{\code{potassium}}{Mass absorption coefficient of potassium (cm2/g).}
\item{\code{calcium}}{Mass absorption coefficient of calcium (cm2/g).}
\item{\code{titanum}}{Mass absorption coefficient of titanum (cm2/g).}
\item{\code{manganese}}{Mass absorption coefficient of manganese (cm2/g).}
\item{\code{iron}}{Mass absorption coefficient of iron (cm2/g).}
}
}
\usage{
ref_mac
}
\description{
Mass Absorption Coefficient
}
\seealso{
Other references: 
\code{\link{interferences}},
\code{\link{isotopes}}
}
\concept{references}
\keyword{datasets}
