% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\docType{data}
\name{std_activity}
\alias{std_activity}
\title{Characteristics of Standard Materials}
\format{
A \code{\link{data.frame}} with 2 rows and 14 variables:
\describe{
\item{\code{standard}}{Name of the standard.}
\item{\code{thickness}}{Thickness (cm).}
\item{\code{volume}}{Volume (cm3).}
\item{\code{mass}}{Mass (g).}
\item{\code{activity_U}}{Uranium specific activity (Bq/kg).}
\item{\code{activity_U_error}}{Uranium specific activity error (Bq/kg).}
\item{\code{activity_Th}}{Thorium specific activity (Bq/kg).}
\item{\code{activity_Th_error}}{Thorium specific activity error (Bq/kg).}
\item{\code{activity_K}}{Potassium specific activity (Bq/kg).}
\item{\code{activity_K_error}}{Potassium specific activity error (Bq/kg).}
\item{\code{content_U}}{Uranium content (ppm).}
\item{\code{content_Th}}{Thorium content (ppm).}
\item{\code{content_K}}{Potassium content (\%).}
}
}
\usage{
std_activity
}
\description{
Characteristics of Standard Materials
}
\seealso{
Other standards: 
\code{\link{clermont}}
}
\concept{standards}
\keyword{datasets}
