% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdXdata-package.R
\docType{package}
\name{bdXdata-package}
\alias{bdXdata-package}
\alias{_PACKAGE}
\alias{bdXdata}
\title{bdXdata: Shared Data in Use at the CRP2A Laboratory}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A collection of shared data in use at the CRP2A laboratory
    (UMR 5060 - CNRS, Université Bordeaux Montaigne). This package
    contains data specific to the laboratory instruments (calibration,
    monitoring, quality control, etc.). These data sets are openly
    distributed in order to ensure transparency and reproducibility of the
    results published by the laboratory team.
}
\details{
\tabular{ll}{
\strong{Package:} \tab bdXdata \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.1.0 \cr
\strong{License:} \tab GPL-3
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://crp2a.github.io/bdXdata/}
  \item \url{https://github.com/crp2a/bdXdata}
  \item Report bugs at \url{https://github.com/crp2a/bdXdata/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Brice Lebrun\cr
\email{brice.lebrun@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
