% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bkg_intensity_BEGe_12cc}
\alias{bkg_intensity_BEGe_12cc}
\title{BEGe 12cc calibration data}
\format{
A list of data frame with 8 variables:
 \describe{
  \item{isotope}{Measured isotope}
  \item{serie}{Radioactive decay series to which the considered isotope belongs.}
  \item{energy}{Theoretical energy of the considered peak.}
  \item{bkg}{Background intensity value (c/ks).}
  \item{bkg_error}{Background error intensity value (c/ks).}
  \item{standard_name}{Name of the used standard.}
  \item{standard}{Standard intensity value (c/ks).}
  \item{standard_error}{Standard error intensity value (c/ks).}
 }
}
\usage{
bkg_intensity_BEGe_12cc
}
\description{
Background and standard intensity values for 12cc BEGe.
}
\seealso{
Other BEGe: 
\code{\link{bkg_intensity_BEGe_60cc}},
\code{\link{standard_characteristics}}
}
\concept{BEGe}
\keyword{datasets}
