% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{standard_characteristics}
\alias{standard_characteristics}
\title{standard characteristics}
\format{
A data frame with 2 rows and 14 variables
 \describe{
  \item{standard}{Name of the standard.}
  \item{thickness}{Thickness in centimeter.}
  \item{volume}{Volume in cubic centimeter.}
  \item{mass}{Mass in grams.}
  \item{activity_U}{Uranium specific activity (Bq/kg).}
  \item{activity_U_error}{Uranium specific activity error (Bq/kg).}
  \item{activity_Th}{Thorium specific activity (Bq/kg).}
  \item{activity_Th_error}{Thorium specific activity error (Bq/kg).}
  \item{activity_K}{Potassium specific activity (Bq/kg).}
  \item{activity_K_error}{Potassium specific activity error (Bq/kg).}
  \item{content_U}{Uranium content (ppm)}
  \item{content_Th}{Thorium content (ppm)}
  \item{content_K}{Potassium content (\%)}
 }
}
\usage{
standard_characteristics
}
\description{
standard characteristics
}
\seealso{
Other BEGe: 
\code{\link{bkg_intensity_BEGe_12cc}},
\code{\link{bkg_intensity_BEGe_60cc}}
}
\concept{BEGe}
\keyword{datasets}
